Update synchronization code
EntityJoinWorldEvent is the best place to send the synchronization
packet from, as it fires after joining the world the first time,
respawning, and changing dimensions.

Previously removed EntityJoinWorldEvent listener was thus reinstated, as
it is very useful.

In 1.8, however, the client player is usually null during the first
login, so added code to handle those cases.